requirements.txt
llama-cpp-python
requests
Flask
Flask-Login
Werkzeug
joblib